he sacado el código de la función de verificación de OTP y lo he metido en un archivo aparte para que sea más fácil de testear y de mantener. También he añadido un reset de la variable OTP en el caso de que el usuario falle en la verificación.
