+++ BACK/controllers/auth.js
@@ -19,7 +19,8 @@ const {userExists} = require('../utils/handleAuth');
 //Importo la funcion para enviar el mail
 const {sendMail} = require('../utils/handleMail');
 
-
+//importo las funciones que manejan la verificacion con OTP
+const {sendOTPVerificationMail, verifyOTP} = require('../utils/handleOTPVerification');
 
 
 const user = require('../models/mongo/user');
@@ -35,7 +36,6 @@ const register = async (req, res) => {
             handleHttpError(res, "ERR_USER_EXISTS", 400);
        }
        else{ 
-        console.log(req.mail);
         if(!await sendOTPVerificationMail(req.mail, res)){
             handleHttpError(res, "ERR_SEND_OTP", 500);
             return;
@@ -96,71 +96,93 @@ const login = async (req, res) => {
 
 
 //Funcion para enviar el mail de verificacion
-const sendOTPVerificationMail = async (email, res) => {
-    try{
-        const otp = Math.floor(1000 + Math.random() * 9000);
+// const sendOTPVerificationMail = async (email, res) => {
+//     try{
+//         const otp = Math.floor(1000 + Math.random() * 9000);
         
-        //Aqui se define la configuracion del mail que se va a enviar
-        const mailOptions = {
-            from: process.env.EMAIL,
-            to: email,
-            subject: "Verificacion de mail",
-           html: `<h3>Estees tu codigo de verificacion:</h3><br><br><h1><b>${otp}</b></h1> <br><br> <p>Este codigo expira en 5 minutos</p>`,
-        };
+//         //Aqui se define la configuracion del mail que se va a enviar
+//         const mailOptions = {
+//             from: process.env.EMAIL,
+//             to: email,
+//             subject: "Verificacion de mail",
+//            html: `<h3>Estees tu codigo de verificacion:</h3><br><br><h1><b>${otp}</b></h1> <br><br> <p>Este codigo expira en 5 minutos</p>`,
+//         };
 
-        //Se hashea el otp
-        const hashedOTP = await encrypt(otp.toString());
-        //Se crea el objeto que se va a guardar en la base de datos
-             const newOtpVerification = await otpVerificationUserModel.create({
-            mail: email,
-            otp: hashedOTP,
-            expires: new Date(new Date().getTime() + 5 * 60 * 1000) // 5 minutes
-        })
-        console.log(newOtpVerification);
-        //Se envia el mail
-        await sendMail(mailOptions);
-        return true;
-    }
-    catch(err){
-        handleHttpError(err, "ERR_SEND_OTP");
-        return false;
-    }
-};
-
-
+//         //Se hashea el otp
+//         const hashedOTP = await encrypt(otp.toString());
+//         //Se crea el objeto que se va a guardar en la base de datos
+//              const newOtpVerification = await otpVerificationUserModel.create({
+//             mail: email,
+//             otp: hashedOTP,
+//             expires: new Date(new Date().getTime() + 5 * 60 * 1000) // 5 minutes
+//         })
+//         console.log(newOtpVerification);
+//         //Se envia el mail
+//         await sendMail(mailOptions);
+//         return true;
+//     }
+//     catch(err){
+//         handleHttpError(err, "ERR_SEND_OTP");
+//         return false;
+//     }
+// };
 
 
-const verifyOTP = async (req, res) => {
+
+
+// const verifyOTP = async (req, res) => {
+//     try{
+//         req = matchedData(req);
+//         const {mail, otp} = req;
+//         console.log(mail);
+//         const userOTP = await otpVerificationUserModel.findOne({mail: mail});
+//         if(!userOTP){
+//             handleHttpError(res, "ERR_OTP_DONT_EXIST", 400);
+//             return false;
+//         }
+
+//         if(new Date() > userOTP.expires){
+//             //await otpVerificationUserModel.deleteMany({mail: mail});
+//             handleHttpError(res, "ERR_OTP_EXPIRED", 400);
+//             return false;
+//          }
+        
+//         const valid = await compare(otp, userOTP.otp);
+//         if(!valid){
+//             handleHttpError(res, "ERR_OTP_INVALID", 400);
+//             return false;
+//         }
+//         await otpVerificationUserModel.deleteOne({mail: mail});
+//         await userModel.updateOne({mail: mail}, {verified: true});
+//         res.send("Email verificado correctamente");
+//         return true;
+//     }
+//     catch(err){
+//         return false;
+//     }
+// }
+
+
+const resendOTP = async (req, res) => {
     try{
         req = matchedData(req);
-        const {mail, otp} = req;
-        console.log(mail);
-        const userOTP = await otpVerificationUserModel.findOne({mail: mail});
-        if(!userOTP){
-            handleHttpError(res, "ERR_OTP_DONT_EXIST", 400);
+        const {mail} = req;
+        //Revisamos si el mail ya esta verificado
+        const user = await userModel.findOne({mail: mail});
+        if(user && user.verified){
+            handleHttpError(res, "ERR_MAIL_VERIFIED", 400);
             return false;
+        } 
+        if(await sendOTPVerificationMail(mail, res)){
+            res.send("Mail sent");
         }
-
-        if(new Date() > userOTP.expires){
-            //await otpVerificationUserModel.deleteMany({mail: mail});
-            handleHttpError(res, "ERR_OTP_EXPIRED", 400);
-            return false;
-         }
-        
-        const valid = await compare(otp, userOTP.otp);
-        if(!valid){
-            handleHttpError(res, "ERR_OTP_INVALID", 400);
-            return false;
-        }
-        await otpVerificationUserModel.deleteOne({mail: mail});
-        await userModel.updateOne({mail: mail}, {verified: true});
-        res.send("Email verificado correctamente");
-        return true;
+        else{
+            res.send("Mail not sent");
+        }  
     }
     catch(err){
-        return false;
+        handleHttpError(res, "ERR_RESEND_OTP");
     }
 }
 
-
-module.exports = {register, login, verifyOTP}
+module.exports = {register, login, resendOTP}
